Simplify generator loop (#7)
Text generation is fast enough with 4k byte blocks that I can consider
it non-blocking, therefore remove the `spawn_blocking` and just make it
all part of the task loop. This also reduces the amount of channels
needing to be created and simplifies how backpressure is applied to the
generation loop.