Input files configuration for generation options (#11)
This adds the ability to configure where input texts for training the
markov chains are located. The configuration allows for many different
chains to be constructed, with the idea of a page having a random chain
granted to it to generate garbage text of different "flavours". The
configuration for this is done with `glob`, so a user can define the
directory structure and text files to load and train.

This means the markov chains are now kept as "state" on the server, and
instead are provided by request handlers on demand.