Route & Socket configuration (#12)
Merged the health route into the main app route so it reflects better
the state of the application (if it returns a response, the main routes
are doing fine), reducing the need for two ports to be used and making
it just one port now. Along with this, I added configuration options for
setting the routes for the pit, so one can have one or multiple
customised routes (to better hide/obfuscate things to malicious
crawlers. They blacklist one route, only to fall for the other). Also
added was configuring the socket address that the server application
listens to for HTTP requests, so to improve the hosting story for
`nailpit`.