The Cratening (#13)
Splits out the main binary project into a main/lib combined package (to
make it easier to write tests for), as well as splitting out a whole
bunch of functionality into smaller crates to reduce the scope of each
part and make it easier to compile for. This is mostly to combat some
bloat on the main crate, but also help to better document what each bit
is responsible for (and also better contain any future unsafe code to
outside the main crate and be contained to smaller, well tested crates
where the API surface can be more strictly enforced).