Use mimalloc for improved performance (#14)
Adds a feature flag (enabled by default) to have `nailpit` make use of
`mimalloc` as the global allocator instead of the system allocator. This
is both a small perf improvement over the system allocator on glibc
Linux, and a much larger improvement over the musl allocator.

To compile `nailpit` to use the system allocator instead, compile with
`--no-default-features`.