Output limiting (#15)
This PR adds some output/rate limiting features to nailpit, this one
focused on adding additional configuration to allow tuning how
quickly/slowly generated output to streamed to the client. This tuning
helps to reduce pressure on the server when generating a lot of content
at once, and keeps clients waiting for more content (as long as the
connection is alive and data being received just often enough to not
disconnect/terminate). The rate is defined by a min and max delay time
in milliseconds, and the delay then is a random value between the two
for each chunk that is delivered. This way, the delay isn't always super
slow, and in theory prevents it looking too uniform/predictable.

The PR includes some extras like configuring the max worker threads
being spawned, as well extra generation options.