Rate Limiting, Part 2 (#16)
Rate Limiting service. This PR introduces custom rate limiting, with the
idea of enabling soft delays (to slow down further a malicious crawler
that is being too insistent on crawling the tarpit) to hard dropping the
request with an appropriate response. This also lays down groundwork for
further features to the rate limiter, such as malicious drop responses
or more advanced delaying tactics.

Included in the PR are ways to configure the rate limiter into various
modes: no limiting, soft limiting, hard limiting and soft with hard
limiting. This can be done with the yml files or environment variables
and is robust enough to ensure that the user will have to do things
*really* wrong for it not to load a config.

The rate limiting service is split out into its own crate for better
encapsulation of functionality and responsibility from the main crate.