Token optimisations (#17)
Makes TokenPair be representable as a faux `u64` type, with optimised
codegen to make use of the fact that it can basically be thrown around,
compared and hashed as a `u64` value instead of two separate `u32`
values. This should improve some throughput metrics for constructing the
markov chains and during generation, but these are expected to be
relatively small.