refactor: separate `nailip`  layer for extracting peer addresses (#19)
Separates out the peer identification process to its own layer, so that
peer info can be shared to the tracing layer for diagnostics and not be
exclusive to the rate limiting service.

Just used `from_fn` for creating the middleware instead of going all out
with `tower::Service` since this is kinda specific to `axum` anyway and
this project.