Poison Prompts in page generation (#21)
Adds the ability to include prompts at the end of a generated page in
order to poison the data further. Add as many prompt strings in the
configuration, and each page will randomly select one to append in the
footer. The list of prompts is also hidden from the logging output so
that it won't spam your terminal/logging infra with the text.