feat: Spicy Rate Limiting (#22)
This PR adds the ability to set up the rate limiting to have a "spicy"
mode, where when the rate limiting hits the "hard drop" level, instead
of dropping the connection with a rate limited status response, it sends
a spicy response instead to attempt to stop the crawler at the source,
so that the crawler is put out of commission or just wastes
resources/money for the crawler.

This mode is only intended for people who know what they are doing and
accept the risks. And naturally, the project will not accept any
liability for misuse or misconfiguration.

## Todos

- [x] Basic Spicy mode, supports brotli encoded responses
- [x] More spicy payload support for different encodings