refactor: Thread per core architecture, custom Stream impl for MarkovGen (#24)
This PR introduces some changes on how `nailpit` works internally in two
major ways: Yeeting the MT `tokio` runtime for a thread-per-core setup
that is seen in `actix-web`, and a custom `Stream` impl for `MarkovGen`
that removes the overhead of maintaining an extra task and a channel for
every response. This turns the response from a push model (a task
pushing bytes into the response) to a pull model (the response pulls
from the stream when ready). This means less allocations and the
hand-written `Stream` is much more efficient as a whole.

Some extra benefits are using `FxHasher` for the `Token`/`TokenPair`
hashing since that is optimised for integer hashing, and some code
tidying to make things a bit neater and well organised.

The combination of changes yields a considerable improvement in the
performance of `nailpit`. In tests designed to batter `nailpit` to see
how it handles heavy loads, here are the following results:

Test Parameters:

- 4 Worker Threads
- 500 connections managed by 24 threads for 2 seconds by rewrk

Before(current main):
```
  Latencies:
    Avg      Stdev    Min      Max      
    151.72ms  52.32ms  3.41ms   314.46ms  
  Requests:
    Total:  6180   Req/Sec: 3141.19
  Transfer:
    Total: 812.80 MB Transfer Rate: 413.13 MB/Sec
+ --------------- + --------------- +
|   Percentile    |   Avg Latency   |
+ --------------- + --------------- +
|      99.9%      |    314.04ms     |
|       99%       |    291.12ms     |
|       95%       |    267.71ms     |
|       90%       |    253.32ms     |
|       75%       |    224.10ms     |
|       50%       |    193.35ms     |
+ --------------- + --------------- +
```

After(this PR):
```
  Latencies:
    Avg      Stdev    Min      Max      
    136.85ms  26.35ms  2.98ms   166.13ms  
  Requests:
    Total:  6853   Req/Sec: 3486.03
  Transfer:
    Total: 884.15 MB Transfer Rate: 449.75 MB/Sec
+ --------------- + --------------- +
|   Percentile    |   Avg Latency   |
+ --------------- + --------------- +
|      99.9%      |    166.06ms     |
|       99%       |    165.28ms     |
|       95%       |    164.19ms     |
|       90%       |    163.04ms     |
|       75%       |    158.68ms     |
|       50%       |    153.85ms     |
+ --------------- + --------------- +
```

In general, the thread-per-core scales better with more workers provided
than with the MT runtime, having better throughput and tail latencies.
Minimum latencies might not be as quick, but overall, the application
handles heavy loads better which is important when dealing with abusive
web crawlers. Very slightly worse minimum latencies won't impact "user"
experience because here we aren't dealing with real "users", just bots
who wanna scrape your content.

Some changes that were tried was doing a custom `hyper`/`axum` server
configuration, but tests yielded no significant or noticeable
improvements over the stock `axum::serve` method.

## Todos

- [x] ~~Experiment with adding `tokio-uring` as an optional runtime
backend~~ This is probably better as a follow-up
- [x] Further code/dep cleanups and review