fix: Gen state machine fixes, runtime improvements (#25)
Basically, refactor introduced some bugs, and this PR fixes them. The
generator state machine wasn't working correctly for delayed output, and
now there's some improvements on the output quality as well. The runtime
is also tidied so it is clearer that the threads don't last for the
entire program and do close when the program begins to terminate,
allowing to avoid some cloning here and there.