refactor: Improved tracing and routing handling (#27)
This PR adds two refactors:

1) Better tracing output for OTEL, using more semantic conventions where
possible, and now that it is in better shape to also have fields based
on the response output, I can add more data later on more easily.

2) Better routing. Now the routing should handle not just nested routes
so like "/private" or "/spicy", but also root level "/". This means the
ground work for templating is in place, since the "static" index/warning
page needs to generate entry links on which path the bot is entering in.