perf: Rejig yielding furing generation, routing tweaks, yeet custom connectinfo (#28)
Just some further tweaks to how routing is done to make it better, the
generation handler and the generation loop to yield less (and using
futures-lite `yield_now` as it has better characteristics than the
`tokio`'s version). Yeeting the custom connection info impl because it
is not really useful even for trace data.