feat: Generation templates (#31)
Closes #29 

This PR introduces a simple templating solution that allows the template
to be streamed incrementally along with the non-blocking updates from
the generation loop. The generation loop is now a bit more complicated,
but it so far has not regressed performance!

The templating is pretty simple, static HTML with a few placeholder
types. Placeholders are specified with `{{ }}`, and the different
placeholder kinds are `title`, `initial`, `main`, `extra`, and `footer`.

- `title` is the main title for the page, and is generated once per
page. More than one invocation per template reuses whatever is
generated. If for a warning page, it is whatever is set as the first
line of the message text file.
- `initial` is for an initial set of paragraph content or a generated
page, best used before the `main` to fill out before going into header
-> paragraph content.
- `main` is for either the static content for a warning page, which is
all the paragraphs separated by newlines *after* the first line (which
acts as the title), or it is the bulk of the generated content that
forms a generated page, following a header -> paragraphs structure.
- `extra` is space for any additional static content relevant for a
generated page, such as prompts.
- `footer` is where the document ends with generated links to other
generated pages. Links are only generated at the footer to keep crawlers
from acting until they have consumed the page in full.

There are two page types, warning page, which only allows `title`,
`main` and `footer` placeholders, and generated page which allows all
listed placeholders.