feat: Multi-platform Docker image (#35)
Time to get `nailpit` prepped for being in a deployable state. This PR
introduces a Dockerfile to allow `nailpit` to be built into a container
with various architectures supported. The intent is to have this
deployable to not just server environments, which can be either x86_64
or arm64, but potentially for Raspberry Pis and any emerging RISCV64
platform. The latter platforms will not be the highest performance ones,
but being able to build and run `nailpit` on these will provide good
feedback on just how efficient the tarpit is.

## TODO

- [x] Initial Dockerfile
- [x] Fill out README on how to build/deploy `nailpit`.
- [x] Document configuration and what is needed to run `nailpit`.