refactor: Switch to actix-web (#37)
A big change! But it seems that my curiosity on whether `actix-web` was
a better fit for `nailpit` was a correct hunch. The end result is the
removal of a bunch of things now handled by `actix-web`, and we get more
performance all-round. We go from with Axum on the generated page
benchmark I set up:

```
Benchmarking 200 connections @ http://0.0.0.0:3000/private/aaaabbbb for 12 second(s)
  Latencies:
    Avg      Stdev    Min      Max      
    33.28ms  5.31ms   1.04ms   56.63ms  
  Requests:
    Total:  71905  Req/Sec: 5994.44
  Transfer:
    Total: 4.83 GB Transfer Rate: 412.03 MB/Sec
+ --------------- + --------------- +
|   Percentile    |   Avg Latency   |
+ --------------- + --------------- +
|      99.9%      |     52.14ms     |
|       99%       |     48.69ms     |
|       95%       |     45.31ms     |
|       90%       |     43.45ms     |
|       75%       |     40.45ms     |
|       50%       |     37.55ms     |
+ --------------- + --------------- +
```

To this with Actix-web:
```
Benchmarking 200 connections @ http://0.0.0.0:3000/private/aaaabbbb for 12 second(s)
  Latencies:
    Avg      Stdev    Min      Max      
    31.81ms  5.10ms   0.56ms   77.80ms  
  Requests:
    Total:  75213  Req/Sec: 6270.62
  Transfer:
    Total: 5.04 GB Transfer Rate: 430.42 MB/Sec
+ --------------- + --------------- +
|   Percentile    |   Avg Latency   |
+ --------------- + --------------- +
|      99.9%      |     70.52ms     |
|       99%       |     52.20ms     |
|       95%       |     44.43ms     |
|       90%       |     42.28ms     |
|       75%       |     38.16ms     |
|       50%       |     35.35ms     |
+ --------------- + --------------- +
```

The difference is more pronounced with the results for the warning page:
Axum:
```
Benchmarking 200 connections @ http://0.0.0.0:3000/private for 12 second(s)
  Latencies:
    Avg      Stdev    Min      Max      
    0.81ms   0.14ms   0.05ms   22.79ms  
  Requests:
    Total: 2958908 Req/Sec: 246699.61
```
Actix-web:
```
Benchmarking 200 connections @ http://0.0.0.0:3000/private for 12 second(s)
  Latencies:
    Avg      Stdev    Min      Max      
    0.48ms   0.08ms   0.02ms   12.32ms  
  Requests:
    Total: 5031346 Req/Sec: 419389.58
```

Overall, it seems actix-web is a bit nicer and does a few things a
little better for my needs, so it stays.