refactor: Return to axum (v2) (#40)
(Retrying due to git/branch shenanigans)

As the PR suggests, I am reverting the changes to switch to actix-web,
because 1) runtime perf advantages didn't quite pan out as expected, 2)
~~issues with how it handles compressing streams~~ (seems to be a
general rust ecosystem thing), and 3) axum + hyper are just better
maintained. Hyper is actually *more* performant than actix-web with my
runtime setup, once you remove the compression layer, so I'm just going
back. Actix-web is interesting, but ultimately, axum+hyper+tower is
nicer from an ecosystem point of view.

## TODOs

- [x] Port tracing strat,
- [x] Serve via hyper instead of axum
- [ ] ~~Add configuration to enable/disable compression~~

## Out of scope

For this PR, switching to a different async runtime outside of tokio is
outside of the scope. Establishing a baseline is important and laying
the foundations to make the switch means for the follow-up work, it
should proceed smoother.