refactor: Async rate limiting (#41)
Switches the previous synchronous access of the rate limiter HashMap to
use the asynchronous entry API, removing a potential deadlock source.
This means moving the rate limit handling into the response future and
making the poll implementation there even more complicated. But in terms
of performance, it is showing no regressions.